flyway depenency added, view of user profile added

+++ src/main/java/me/lexik/webapp/domain/Message.java
@@ -7,8 +7,8 @@ import javax.persistence.*;
 public class Message {
 
     @Id
-    @GeneratedValue(strategy=GenerationType.SEQUENCE)
-    private Integer id;
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    private Long id;
     private String text;
     private String tag;
 
@@ -27,13 +27,14 @@ public class Message {
     }
 
     public String getAuthorName() {
-        return author!= null ? author.getUsername() : "<none>";
+        return author != null ? author.getUsername() : "<none>";
     }
-    public Integer getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
